oranized delete post hook
